requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.9
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
